refactor: route sliced imports through public APIs
