+++ src/app/auth/logout.spec.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
 import { logout } from "@/app/auth/logout";
-import { STUDY_STORAGE_KEY, studyStore } from "@/features/study/state/studyStore";
+import { studyStore } from "@/features/study";
 
 const mocks = vi.hoisted(() => ({
   signOutCurrentUser: vi.fn(),
@@ -44,7 +44,7 @@ describe("logout", () => {
 
     expect(operations).toEqual(["suspend", "sign-out", "stop-remote", "resume"]);
     expect(studyStore.getState().sessionsByDeckId).toEqual({});
-    expect(localStorage.getItem(STUDY_STORAGE_KEY)).toBeNull();
+    expect(localStorage).toHaveLength(0);
   });
 
   it("preserves local state when sign-out fails", async () => {
